Refactor exec/expression into exec/call and exec/new (#529)
