+++ boa/src/exec/call/mod.rs
@@ -1,12 +1,7 @@
-//! Expression execution.
-
 use super::{Executable, Interpreter, InterpreterState};
 use crate::{
-    builtins::{
-        object::{ObjectData, INSTANCE_PROTOTYPE, PROTOTYPE},
-        value::{ResultValue, Type, Value},
-    },
-    syntax::ast::node::{Call, New, Node},
+    builtins::value::{ResultValue, Type},
+    syntax::ast::node::{Call, Node},
     BoaProfiler,
 };
 
@@ -51,34 +46,5 @@ impl Executable for Call {
         interpreter.set_current_state(InterpreterState::Executing);
 
         fnct_result
-    }
-}
-
-impl Executable for New {
-    fn run(&self, interpreter: &mut Interpreter) -> ResultValue {
-        // let (callee, args) = match call.as_ref() {
-        //     Node::Call(callee, args) => (callee, args),
-        //     _ => unreachable!("Node::New(ref call): 'call' must only be Node::Call type."),
-        // };
-
-        let func_object = self.expr().run(interpreter)?;
-        let mut v_args = Vec::with_capacity(self.args().len());
-        for arg in self.args() {
-            v_args.push(arg.run(interpreter)?);
-        }
-        let this = Value::new_object(None);
-        // Create a blank object, then set its __proto__ property to the [Constructor].prototype
-        this.set_internal_slot(INSTANCE_PROTOTYPE, func_object.get_field(PROTOTYPE));
-
-        match func_object {
-            Value::Object(ref obj) => {
-                let obj = obj.borrow();
-                if let ObjectData::Function(ref func) = obj.data {
-                    return func.construct(func_object.clone(), &this, &v_args, interpreter);
-                }
-                interpreter.throw_type_error("not a constructor")
-            }
-            _ => Ok(Value::undefined()),
-        }
     }
 }
