+++ boa/src/exec/mod.rs
@@ -3,13 +3,14 @@
 mod array;
 mod block;
 mod break_node;
+mod call;
 mod conditional;
 mod declaration;
 mod exception;
-mod expression;
 mod field;
 mod identifier;
 mod iteration;
+mod new;
 mod object;
 mod operator;
 mod return_smt;
@@ -630,7 +631,7 @@ impl Executable for Node {
             Node::Identifier(ref identifier) => identifier.run(interpreter),
             Node::GetConstField(ref get_const_field_node) => get_const_field_node.run(interpreter),
             Node::GetField(ref get_field) => get_field.run(interpreter),
-            Node::Call(ref expr) => expr.run(interpreter),
+            Node::Call(ref call) => call.run(interpreter),
             Node::WhileLoop(ref while_loop) => while_loop.run(interpreter),
             Node::DoWhileLoop(ref do_while) => do_while.run(interpreter),
             Node::ForLoop(ref for_loop) => for_loop.run(interpreter),
@@ -641,7 +642,7 @@ impl Executable for Node {
             // <https://tc39.es/ecma262/#sec-createdynamicfunction>
             Node::FunctionDecl(ref decl) => decl.run(interpreter),
             // <https://tc39.es/ecma262/#sec-createdynamicfunction>
-            Node::FunctionExpr(ref expr) => expr.run(interpreter),
+            Node::FunctionExpr(ref function_expr) => function_expr.run(interpreter),
             Node::ArrowFunctionDecl(ref decl) => decl.run(interpreter),
             Node::BinOp(ref op) => op.run(interpreter),
             Node::UnaryOp(ref op) => op.run(interpreter),
